quickly implement index page / new form distinction for requests
